Moved Hangfire server to main command line
