--- src/JobServer/JobServer.csproj
@@ -1,31 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <OutputType>Exe</OutputType>
-        <TargetFramework>net6.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <RootNamespace>Scrap.JobServer</RootNamespace>
-        <IsPackable>false</IsPackable>
-        <ImplicitUsings>true</ImplicitUsings>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <PackageReference Include="Figgle" Version="0.4.0" />
-      <PackageReference Include="Hangfire.SqlServer" Version="1.7.28" />
-      <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.0.1" />
-      <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="6.0.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\DependencyInjection\DependencyInjection.csproj" />
-      <ProjectReference Include="..\Domain\Domain.csproj" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Update="appsettings.json">
-        <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-      </None>
-    </ItemGroup>
-
-</Project>
