--- src/JobServer/appsettings.json
@@ -1,17 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Trace",
-      "System": "Warning",
-      "Microsoft": "Warning",
-      "Hangfire": "Information"
-    }
-  },
-  "Hangfire": {
-    "Database": "Data Source=(localdb)\\MSSQLLocalDB"
-  },
-  "Scrap": {
-    "Database": "Filename=C:\\Users\\ignacio.calvo\\Dropbox\\Ignacio\\Scrap\\scrap.db;Connection=shared",
-    "Definitions": "C:\\Users\\ignacio.calvo\\Dropbox\\Ignacio\\Scrap\\jobDefinitions.json"
-  }
-}
